pyproject.toml
[project]
name = "self-learning-viral-agent"
version = "0.1.0"
description = "Autonomous AI agent that grows your Threads account through self-learning content loops — researches viral posts, generates content, publishes, measures, and adapts strategy."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = ["langgraph", "ai-agent", "threads", "social-media", "autonomous", "self-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "langgraph>=0.3",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langgraph-checkpoint-postgres>=2.0",
    "psycopg[binary]>=3.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "praw>=7.7",
    "apify-client>=1.6",
    "apscheduler>=3.10",
    "python-telegram-bot>=21.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "langsmith>=0.1",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
